test: share SOAP toolkit factory test helpers
