+++ toolkit/src/test/kotlin/com/izivia/ocpp/integration/test/Ocpp15FactoryTest.kt
@@ -7,19 +7,9 @@ import com.izivia.ocpp.integration.ApiFactory
 import com.izivia.ocpp.integration.model.Settings
 import com.izivia.ocpp.integration.model.TransportEnum
 import com.izivia.ocpp.operation.information.RequestMetadata
-import com.izivia.ocpp.soap.OcppSoapParser
-import com.izivia.ocpp.soap.ResponseSoapMessage
 import com.izivia.ocpp.soap15.Ocpp15SoapParser
 import com.izivia.ocpp.transport.OcppVersion
 import io.mockk.mockk
-import kotlinx.datetime.Instant
-import org.http4k.core.Method
-import org.http4k.core.Response
-import org.http4k.core.Status
-import org.http4k.routing.bind
-import org.http4k.routing.routes
-import org.http4k.server.Undertow
-import org.http4k.server.asServer
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
 import strikt.assertions.isA
@@ -27,9 +17,6 @@ import strikt.assertions.isEqualTo
 import com.izivia.ocpp.core15.model.heartbeat.HeartbeatResp as HeartbeatResp15
 
 class Ocpp15FactoryTest {
-    private val messageId = "a727d144-82bb-497a-a0c7-4ef2295910d4"
-    private val currentTime = Instant.parse("2022-02-15T00:00:00.000Z")
-
     @Test
     fun `creates generic API adapter for OCPP 1-5`() {
         val api = ApiFactory.getCSMSApi(
@@ -51,17 +38,17 @@ class Ocpp15FactoryTest {
         val receivedActions = mutableListOf<String>()
         val server = soapServer(
             parser = Ocpp15SoapParser(),
-            payload = HeartbeatResp15(currentTime),
+            payload = HeartbeatResp15(TEST_CURRENT_TIME),
             receivedActions = receivedActions
         )
 
         try {
-            val api = soapApi(server.port())
+            val api = soapApi(OcppVersion.OCPP_1_5, server.port())
             try {
                 api.connect()
                 val response = api.heartbeat(RequestMetadata("CP001"), HeartbeatReq())
 
-                expectThat(response.response.currentTime).isEqualTo(currentTime)
+                expectThat(response.response.currentTime).isEqualTo(TEST_CURRENT_TIME)
                 expectThat(receivedActions.toList()).isEqualTo(listOf("Heartbeat"))
             } finally {
                 api.close()
@@ -70,42 +57,4 @@ class Ocpp15FactoryTest {
             server.close()
         }
     }
-
-    private fun soapApi(port: Int) =
-        ApiFactory.getCSMSApi(
-            settings = Settings(
-                ocppVersion = OcppVersion.OCPP_1_5,
-                transportType = TransportEnum.SOAP,
-                target = "http://localhost:$port/ocpp",
-                clientPath = "/cp",
-                clientPort = 0,
-                newMessageId = { messageId }
-            ),
-            ocppId = "CP001",
-            csApi = mockk<CSApi>(relaxed = true)
-        )
-
-    private fun soapServer(
-        parser: OcppSoapParser,
-        payload: Any,
-        receivedActions: MutableList<String>
-    ) = routes(
-        "/ocpp/" bind Method.POST to { request ->
-            val soapRequest = parser.parseAnyRequestFromSoap(request.bodyString())
-            receivedActions += soapRequest.action
-            Response(Status.OK).body(
-                parser.mapResponseToSoap(
-                    ResponseSoapMessage(
-                        action = soapRequest.action,
-                        messageId = "response-$messageId",
-                        relatesTo = soapRequest.messageId,
-                        to = soapRequest.from,
-                        from = soapRequest.to,
-                        chargeBoxIdentity = soapRequest.chargingStationId,
-                        payload = payload
-                    )
-                )
-            )
-        }
-    ).asServer(Undertow(0)).start()
 }
